refactor: move internal/jobs under internal/app
