+++ internal/app/jobs/remove.go
